requirements.txt
astroid==2.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-cors-headers==3.1.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
numpy==1.17.2
pandas==0.25.1
psycopg2==2.8.3
pylint==2.4.1
python-dateutil==2.8.0
pytz==2019.2
ratelimit==2.2.1
requests==2.22.0
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
